fix: 難易度設定を別ページで表示
